Cleanup casting in page builder
Make the entire class expect the same size types since the pak file format was designed around this. Just limit casting to CPakFile::CreatePageLump() because that function is used in asset code and there we work with 64 bit size types to avoid putting casts everywhere in code.

+++ src/logic/pakfile.cpp
@@ -369,7 +369,7 @@ void CPakFile::GenerateGuidData()
 
 PakPageLump_s CPakFile::CreatePageLump(const size_t size, const int flags, const int alignment, void* const buf)
 {
-	return m_pageBuilder.CreatePageLump(size, flags, alignment, buf);
+	return m_pageBuilder.CreatePageLump(static_cast<int>(size), flags, alignment, buf);
 }
 
 //-----------------------------------------------------------------------------
